requirements.txt
PyMySQL>=1.0.2
GitPython>=3.1.30
PyYAML>=6.0
colorama>=0.4.4
